fix: Facility 섹션 분리
